api/requirements.txt
# 🚀 Performance Optimized (Vercel-Safe)
fastapi==0.111.0
uvicorn==0.30.1
python-multipart==0.0.9
python-dotenv==1.0.1
langchain==0.2.1
langchain-core==0.2.17
langchain-community==0.2.1
langchain-google-genai==1.0.8
unstructured-client==0.24.1
pydantic
pydantic-settings
google-generativeai==0.7.2
requests==2.31.0
lxml==5.2.2
# ⚠️ Removed ChromaDB (Too heavy for Vercel/Hobby)